# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "Users_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/v1/teams": {
      "get": {
        "operationId": "Teams_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "region",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tournament_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "created_at"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "teams"
        ]
      }
    },
    "/api/v1/teams/{id}/statistics": {
      "get": {
        "operationId": "Teams_statistics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "teams"
        ]
      }
    },
    "/api/v1/teams/{id}/roster": {
      "get": {
        "operationId": "Teams_roster",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "teams"
        ]
      }
    },
    "/api/v1/teams/{id}/matches": {
      "get": {
        "operationId": "Teams_teamMatches",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "scheduled",
                "live",
                "completed",
                "cancelled",
                "postponed"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "teams"
        ]
      }
    },
    "/api/v1/teams/{id}": {
      "get": {
        "operationId": "Teams_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "teams"
        ]
      }
    },
...
    "/api/v1/matches": {
      "get": {
        "operationId": "Matches_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "scheduled",
                "live",
                "completed",
                "cancelled",
                "postponed"
              ]
            }
          },
          {
            "name": "tournament_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "team_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featured",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "scheduled_at",
                "viewer_count"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
...
    "/api/v1/matches/{id}/ticket": {
      "get": {
        "operationId": "MatchTickets_availability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/v1/matches/{id}/orders": {
      "post": {
        "operationId": "MatchTickets_createOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "matches"
        ]
      }
    },
    "/api/v1/me/orders": {
      "get": {
        "operationId": "MeTickets_listOrders",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "cancelled",
                "pending",
                "paid",
                "expired",
                "failed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/orders/{id}": {
      "get": {
        "operationId": "MeTickets_getOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/orders/{id}/cancel": {
      "post": {
        "operationId": "MeTickets_cancelOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/tickets": {
      "get": {
        "operationId": "MeTickets_listTickets",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "match_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/admin/matches/{id}/ticket-config": {
      "put": {
        "operationId": "AdminTickets_upsertConfig",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/orders": {
      "get": {
        "operationId": "AdminTickets_listOrders",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "match_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "cancelled",
                "pending",
                "paid",
                "expired",
                "failed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/webhooks/nowpayments": {
      "post": {
        "operationId": "PaymentWebhooks_handleNowPayments",
        "parameters": [
          {
            "name": "x-nowpayments-sig",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "webhooks"
        ]
      }
    },
    "/api/v1/admin/matches": {
      "post": {
        "operationId": "Admin_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "THUMBZ API",
    "description": "Mobile Legends esports catalog, live match data, comments, ticketing and payments. Human-written rules and rationale live in docs/API-CONTRACT.md.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase access token"
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 4
          }
        },
        "required": [
          "quantity"
        ]
      },
      "TicketConfigDto": {
        "type": "object",
        "properties": {
          "venue_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "venue_city": {
            "type": "string",
            "maxLength": 120
          },
          "price_usd": {
            "type": "number",
            "minimum": 0.01
          },
          "quota_total": {
            "type": "number",
            "minimum": 1
          },
          "sales_open_at": {
            "type": "string"
          },
          "sales_close_at": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          }
        },
        "required": [
          "venue_name",
          "price_usd",
          "quota_total"
        ]
      },
      "CreateMatchDto": {
        "type": "object",
        "properties": {
          "tournament_id": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "stage": {
            "type": "string",
            "nullable": true,
            "enum": [
              "group_stage",
              "regular_season",
              "playoffs",
              "semifinal",
              "third_place",
              "grand_final"
            ]
          },
          "round": {
            "type": "number",
            "nullable": true
          },
          "group_name": {
            "type": "string",
            "nullable": true
          },
          "best_of": {
            "type": "number",
            "minimum": 1
          },
          "game_number": {
            "type": "number",
            "nullable": true
          },
          "team_a_id": {
            "type": "string",
            "format": "uuid"
          },
          "team_b_id": {
            "type": "string",
            "format": "uuid"
          },
          "score_a": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "score_b": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "winner_team_id": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "live",
              "completed",
              "cancelled",
              "postponed"
            ]
          },
          "scheduled_at": {
            "type": "string"
          },
          "started_at": {
            "type": "string",
            "nullable": true
          },
          "ended_at": {
            "type": "string",
            "nullable": true
          },
          "stream_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "thumbnail_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "viewer_count": {
            "type": "number",
            "minimum": 0
          },
          "featured": {
            "type": "boolean"
          }
        },
        "required": [
          "team_a_id",
          "team_b_id",
          "scheduled_at"
        ]
      },
...
    }
  }
}